feat(edit) : 에러 처리 통합, 스켈레톤 적용
