+++ src/app/owner/my-shop/[shopId]/shop-edit/page.tsx
@@ -1,4 +1,4 @@
-import ShopEditClient from './ShopEditClient';
+import ShopEditClient from './shop-edit-client';
 
 export default function ShopEditPage() {
   return <ShopEditClient />;
